python: Fixed #982, moved GDB printers to a separate toplevel folder.
Other non GDB related inet and omnetpp python packages are not imported anymore.
