+++ .gdbinit.py
@@ -7,6 +7,6 @@ sys.path.insert(0, os.path.dirname(os.path.abspath(__file__))+"/misc/gdb")
 if 'register_inet_printers' in dir():
     print('inet pretty printers already initialized.')
 else:
-    from inet.printers import register_inet_printers
+    from inetgdb.printers import register_inet_printers
     register_inet_printers(None)
     print('Pretty printers initialized: inet')
